requirements.txt
flet
pyshorteners
pystray
Pillow